% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertxlsform.R
\name{convertxlsform}
\alias{convertxlsform}
\title{Convert ODK XLSForms to Word Documents}
\usage{
convertxlsform(xlsform_path, selected_language = "en", output_docx = NULL)
}
\arguments{
\item{xlsform_path}{Character. Path to the XLSForm file (in `.xlsx` format).}

\item{selected_language}{Character. The language to use for label and hint columns. Default is `"en"`.}

\item{output_docx}{Character. The output Word document filename. If `NULL`, the function uses the XLSForm’s base name appended with `"_PAPER.docx"`.}
}
\value{
Invisibly returns the modified Word document object. The DOCX file is written to disk.
}
\description{
The `convertxlsform` function takes an XLSForm questionnaire from ODK (Open Data Kit)
and converts it into a neatly formatted Microsoft Word document.
It supports multiple languages (if language-specific columns exist) and applies various formatting
options such as bolding required questions, italicizing hints and constraint messages, and more.
}
\details{
If the XLSForm does not have language-specific columns (i.e. columns named like `label::<lang>`),
the function will use the default columns and will not print the "Selected language" line.
}
\examples{
\dontrun{
  convertxlsform("path/to/your/form.xlsx", selected_language = "English (en)")
}

}
